+++ src/index.ts
@@ -17,6 +17,7 @@ import { subscriptionSteps } from './steps/resource-manager/subscriptions';
 import { apiManagementSteps } from './steps/resource-manager/api-management';
 import { dnsSteps } from './steps/resource-manager/dns';
 import { privateDnsSteps } from './steps/resource-manager/private-dns';
+import { containerRegistrySteps } from './steps/resource-manager/container-registry';
 
 export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = {
   instanceConfigFields: {
@@ -60,5 +61,6 @@ export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = 
     ...apiManagementSteps,
     ...dnsSteps,
     ...privateDnsSteps,
+    ...containerRegistrySteps,
   ],
 };